Discord関係の処理を discord.go に分離
